requirements.txt
# Creative AI Shorts Platform

fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-multipart>=0.0.6

# Existing StoryGenius dependencies
moviepy>=2.0.0
edge-tts>=6.1.0
google-cloud-aiplatform>=1.38.0
Pillow>=10.0.0
